# public/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"IInstitution": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"institutionType": {
						"type": "string"
					},
					"level": {
						"type": "string"
					},
					"hasPharmacy": {
						"type": "boolean"
					},
					"admin": {
						"properties": {
							"email": {
								"type": "string"
							},
							"phone": {
								"type": "string"
							},
							"name": {
								"type": "string"
							}
						},
						"required": [
							"email",
							"phone",
							"name"
						],
						"type": "object"
					},
					"details": {
						"properties": {
							"percentage": {
								"type": "number",
								"format": "double"
							},
							"TIN": {
								"type": "string"
							},
							"location": {
								"type": "string"
							}
						},
						"required": [
							"percentage",
							"TIN",
							"location"
						],
						"type": "object"
					},
					"institutionId": {
						"type": "string",
						"nullable": true
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"name",
					"institutionType",
					"level",
					"hasPharmacy",
					"admin",
					"details",
					"createdAt"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"IServiceAct": {
				"properties": {
					"id": {
						"type": "string"
					},
					"label": {
						"type": "string"
					},
					"desc": {
						"type": "string"
					},
					"price": {
						"type": "number",
						"format": "double"
					},
					"serviceId": {
						"type": "string"
					},
					"institutionId": {
						"type": "string"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"service": {
						"$ref": "#/components/schemas/IService"
					},
					"institution": {
						"$ref": "#/components/schemas/IInstitution"
					}
				},
				"required": [
					"id",
					"label",
					"desc",
					"price",
					"serviceId",
					"institutionId",
					"createdAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IService": {
				"properties": {
					"id": {
						"type": "string"
					},
					"label": {
						"type": "string"
					},
					"desc": {
						"type": "string"
					},
					"level": {
						"type": "string"
					},
					"assignDuringOrientation": {
						"type": "boolean"
					},
					"institutionId": {
						"type": "string"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"acts": {
						"items": {
							"$ref": "#/components/schemas/IServiceAct"
						},
						"type": "array"
					},
					"institution": {
						"$ref": "#/components/schemas/IInstitution"
					}
				},
				"required": [
					"id",
					"label",
					"desc",
					"level",
					"assignDuringOrientation",
					"institutionId",
					"createdAt"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"IPaged_IServiceResponse_": {
				"properties": {
					"data": {
						"$ref": "#/components/schemas/IServiceResponse"
					},
					"totalItems": {
						"type": "number",
						"format": "double"
					},
					"currentPage": {
						"type": "number",
						"format": "double"
					},
					"itemsPerPage": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"data",
					"totalItems",
					"currentPage",
					"itemsPerPage"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"IServiceRequest": {
				"properties": {
					"label": {
						"type": "string"
					},
					"institutionId": {
						"type": "string"
					},
					"desc": {
						"type": "string"
					},
					"institution": {
						"$ref": "#/components/schemas/IInstitution"
					},
					"level": {
						"type": "string"
					},
					"assignDuringOrientation": {
						"type": "boolean"
					},
					"acts": {
						"items": {
							"$ref": "#/components/schemas/IServiceAct"
						},
						"type": "array"
					},
					"id": {
						"type": "string"
					}
				},
				"required": [
					"label",
					"institutionId",
					"desc",
					"level",
					"assignDuringOrientation"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"IInstitutionDTO": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"institutionType": {
						"type": "string"
					},
					"level": {
						"type": "string"
					},
					"hasPharmacy": {
						"type": "boolean"
					},
					"admin": {
						"properties": {
							"email": {
								"type": "string"
							},
							"phone": {
								"type": "string"
							},
							"name": {
								"type": "string"
							}
						},
						"required": [
							"email",
							"phone",
							"name"
						],
						"type": "object"
					},
					"details": {
						"properties": {
							"percentage": {
								"type": "number",
								"format": "double"
							},
							"TIN": {
								"type": "string"
							},
							"location": {
								"type": "string"
							}
						},
						"required": [
							"percentage",
							"TIN",
							"location"
						],
						"type": "object"
					},
					"institutionId": {
						"type": "string",
						"nullable": true
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"users": {
						"items": {
							"$ref": "#/components/schemas/UserReponse"
						},
						"type": "array"
					},
					"parentInstitution": {
						"$ref": "#/components/schemas/IInstitution"
					},
					"branches": {
						"items": {
							"$ref": "#/components/schemas/IInstitution"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"name",
					"institutionType",
					"level",
					"hasPharmacy",
					"admin",
					"details",
					"createdAt"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"IPaged_IInstitutionResponse_": {
				"properties": {
					"data": {
						"$ref": "#/components/schemas/IInstitutionResponse"
					},
					"totalItems": {
						"type": "number",
						"format": "double"
					},
					"currentPage": {
						"type": "number",
						"format": "double"
					},
					"itemsPerPage": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"data",
					"totalItems",
					"currentPage",
					"itemsPerPage"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"IInstitutionRequest": {
				"properties": {
					"name": {
						"type": "string"
					},
					"institutionId": {
						"type": "string"
					},
					"level": {
						"type": "string"
					},
					"admin": {
						"properties": {
							"email": {
								"type": "string"
							},
							"phone": {
								"type": "string"
							},
							"name": {
								"type": "string"
							}
						},
						"required": [
							"email",
							"phone",
							"name"
						],
						"type": "object"
					},
					"details": {
						"properties": {
							"percentage": {
								"type": "number",
								"format": "double"
							},
							"TIN": {
								"type": "string"
							},
							"location": {
								"type": "string"
							}
						},
						"required": [
							"percentage",
							"TIN",
							"location"
						],
						"type": "object"
					},
					"institutionType": {
						"type": "string",
						"nullable": true
					},
					"hasPharmacy": {
						"type": "boolean"
					}
				},
				"required": [
					"name",
					"level",
					"admin",
					"details"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ICreateBranch": {
				"properties": {
					"name": {
						"type": "string"
					},
					"location": {
						"type": "string"
					},
					"details": {
						"properties": {
							"percentage": {
								"type": "number",
								"format": "double"
							},
							"TIN": {
								"type": "string"
							},
							"location": {
								"type": "string"
							}
						},
						"required": [
							"percentage",
							"TIN",
							"location"
						],
						"type": "object"
					}
				},
				"required": [
					"name",
					"location",
					"details"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"IExam": {
				"properties": {
					"id": {
						"type": "string"
					},
					"exam_code": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"price": {
						"type": "number",
						"format": "double"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"institutionExam": {
						"items": {
							"$ref": "#/components/schemas/IInstitutionExam"
						},
						"type": "array"
					},
					"insuranceExam": {
						"items": {
							"$ref": "#/components/schemas/IInsuranceExam"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"exam_code",
					"description",
					"name",
					"price",
					"createdAt"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"IFormDTO": {
				"properties": {
					"id": {
						"type": "string"
					},
					"institutionId": {
						"type": "string"
					},
					"insuranceId": {
						"type": "string",
						"nullable": true
					},
					"insuranceCard": {
						"type": "string",
						"nullable": true
					},
					"patientId": {
						"type": "string"
					},
					"formNO": {
						"type": "string"
					},
					"at": {
						"type": "string"
					},
					"from": {
						"type": "string"
					},
					"isOpen": {
						"type": "boolean"
					},
					"details": {
						"properties": {
							"height": {
								"type": "number",
								"format": "double",
								"nullable": true
							},
							"weight": {
								"type": "number",
								"format": "double",
								"nullable": true
							},
							"temperature": {
								"type": "number",
								"format": "double",
								"nullable": true
							}
						},
						"required": [
							"height",
							"weight",
							"temperature"
						],
						"type": "object"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"patient": {
						"$ref": "#/components/schemas/IPatient"
					},
					"institution": {
						"$ref": "#/components/schemas/IInstitution"
					},
					"insurance": {
						"$ref": "#/components/schemas/IInstitution"
					},
					"drugs": {
						"items": {
							"$ref": "#/components/schemas/IFormDrugDTO"
						},
						"type": "array"
					},
					"exams": {
						"items": {
							"$ref": "#/components/schemas/IFormExamDTO"
						},
						"type": "array"
					},
					"consultations": {
						"items": {
							"$ref": "#/components/schemas/IFormConsultationDTO"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"institutionId",
					"insuranceId",
					"insuranceCard",
					"patientId",
					"formNO",
					"at",
					"from",
					"isOpen",
					"details",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"IFormConsultationRequest": {
				"properties": {
					"exams": {
						"items": {
							"$ref": "#/components/schemas/FormAddExam"
						},
						"type": "array"
					},
					"drugs": {
						"items": {
							"$ref": "#/components/schemas/FormAddDrug"
						},
						"type": "array"
					},
					"verdict": {
						"type": "string"
					},
					"consultationId": {
						"type": "string"
					}
				},
				"required": [
					"exams",
					"drugs",
					"verdict",
					"consultationId"
				],
				"type": "object"
			},
...
			"IInstitutionDrug": {
				"properties": {
					"id": {
						"type": "string"
					},
					"drugId": {
						"type": "string"
					},
					"institutionId": {
						"type": "string"
					},
					"purchaseId": {
						"type": "string"
					},
					"drugPurchaseId": {
						"type": "string"
					},
					"insuranceDrugId": {
						"type": "string",
						"nullable": true
					},
					"batchNumber": {
						"type": "string",
						"nullable": true
					},
					"expireDate": {
						"type": "string",
						"format": "date-time",
						"nullable": true
					},
					"itemNo": {
						"type": "string"
					},
					"quantity": {
						"type": "number",
						"format": "double"
					},
					"price": {
						"type": "number",
						"format": "double"
					},
					"isAvailable": {
						"type": "boolean"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"deletedAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"totalQuantity": {
						"type": "number",
						"format": "double"
					},
					"drug": {
						"$ref": "#/components/schemas/IDrug"
					},
					"insuranceDrug": {
						"$ref": "#/components/schemas/IInsuranceDrug"
					}
				},
				"required": [
					"id",
					"drugId",
					"institutionId",
					"purchaseId",
					"drugPurchaseId",
					"insuranceDrugId",
					"batchNumber",
					"expireDate",
					"itemNo",
					"quantity",
					"price",
					"isAvailable",
					"createdAt",
					"deletedAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"IDrugCategory": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"name",
					"createdAt"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"IPaged_IInstitutionDrugResponse_": {
				"properties": {
					"data": {
						"$ref": "#/components/schemas/IInstitutionDrugResponse"
					},
					"totalItems": {
						"type": "number",
						"format": "double"
					},
					"currentPage": {
						"type": "number",
						"format": "double"
					},
					"itemsPerPage": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"data",
					"totalItems",
					"currentPage",
					"itemsPerPage"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"jwtAuth": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT",
				"description": "Bearer token to access these API endpoints"
			}
		}
	},
	"info": {
		"title": "digital-med",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {
			"name": "Reponse Ashimwe"
		}
	},
	"paths": {
...
		"/api/service-acts/all": {
			"get": {
				"operationId": "GetNPaged",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/IServiceAct"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Service Acts"
				],
				"security": [
					{
						"jwtAuth": []
					}
				],
				"parameters": []
			}
		},
...
		"/api/services": {
			"get": {
				"operationId": "GetAll",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IPaged_IServiceResponse_"
								}
							}
						}
					}
				},
				"tags": [
					"Services"
				],
				"security": [
					{
						"jwtAuth": []
					}
				],
				"parameters": []
			},
			"post": {
				"operationId": "Create",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IService"
								}
							}
						}
					}
				},
				"tags": [
					"Services"
				],
				"security": [
					{
						"jwtAuth": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IServiceRequest"
							}
						}
					}
				}
			}
		},
		"/api/services/{id}": {
			"get": {
				"operationId": "GetOne",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IService"
								}
							}
						}
					}
				},
				"tags": [
					"Services"
				],
				"security": [
					{
						"jwtAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "Update",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "boolean"
								}
							}
						}
					}
				},
				"tags": [
					"Services"
				],
				"security": [
					{
						"jwtAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IServiceRequest"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "Delete",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "boolean"
								}
							}
						}
					}
				},
				"tags": [
					"Services"
				],
				"security": [
					{
						"jwtAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/services/all": {
			"get": {
				"operationId": "GetNPaged",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/IService"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Services"
				],
				"security": [
					{
						"jwtAuth": []
					}
				],
				"parameters": []
			}
		},
...
		"/api/institutions": {
			"get": {
				"operationId": "GetAll",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IPaged_IInstitutionResponse_"
								}
							}
						}
					}
				},
				"tags": [
					"Institutions"
				],
				"security": [
					{
						"jwtAuth": []
					}
				],
				"parameters": []
			},
			"post": {
				"operationId": "Create",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IInstitution"
								}
							}
						}
					}
				},
				"tags": [
					"Institutions"
				],
				"security": [
					{
						"jwtAuth": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IInstitutionRequest"
							}
						}
					}
				}
			}
		},
		"/api/institutions/{id}": {
			"get": {
				"operationId": "GetOne",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IInstitutionDTO"
								}
							}
						}
					}
				},
				"tags": [
					"Institutions"
				],
				"security": [
					{
						"jwtAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "Update",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "boolean"
								}
							}
						}
					}
				},
				"tags": [
					"Institutions"
				],
				"security": [
					{
						"jwtAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IInstitutionRequest"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "Delete",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "boolean"
								}
							}
						}
					}
				},
				"tags": [
					"Institutions"
				],
				"security": [
					{
						"jwtAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/institutions/pharmacies": {
			"get": {
				"operationId": "GetPharmacies",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/IInstitutionDTO"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Institutions"
				],
				"security": [
					{
						"jwtAuth": []
					}
				],
				"parameters": []
			}
		},
		"/api/institutions/insurances": {
			"get": {
				"operationId": "GetInsurances",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/IInstitutionDTO"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Institutions"
				],
				"security": [
					{
						"jwtAuth": []
					}
				],
				"parameters": []
			}
		},
		"/api/institutions/branches": {
			"post": {
				"operationId": "CreateBranches",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IInstitution"
								}
							}
						}
					}
				},
				"tags": [
					"Institutions"
				],
				"security": [
					{
						"jwtAuth": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ICreateBranch"
							}
						}
					}
				}
			}
		},
		"/api/institutions/all": {
			"get": {
				"operationId": "All",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/IInstitutionDTO"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Institutions"
				],
				"security": [
					{
						"jwtAuth": []
					}
				],
				"parameters": []
			}
		},
...
		"/api/forms/{id}/consultation": {
			"post": {
				"operationId": "Consultation",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IFormDTO"
								}
							}
						}
					}
				},
				"tags": [
					"Forms"
				],
				"security": [
					{
						"jwtAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IFormConsultationRequest"
							}
						}
					}
				}
			}
		},
...
		"/api/exams/all": {
			"get": {
				"operationId": "GetNPaged",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/IExam"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Exams"
				],
				"security": [
					{
						"jwtAuth": []
					}
				],
				"parameters": []
			}
		},
...
		"/api/drug-categories/all": {
			"get": {
				"operationId": "GetNPaged",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/IDrugCategory"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Drugs category"
				],
				"security": [
					{
						"jwtAuth": []
					}
				],
				"parameters": []
			}
		},
...
		"/api/drugs/institution": {
			"get": {
				"operationId": "GetInstitutionDrugs",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IPaged_IInstitutionDrugResponse_"
								}
							}
						}
					}
				},
				"tags": [
					"Drugs"
				],
				"security": [
					{
						"jwtAuth": []
					}
				],
				"parameters": []
			}
		},
...
		"/api/drugs/institution/all": {
			"get": {
				"operationId": "InstitutionAll",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/IInstitutionDrug"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Drugs"
				],
				"security": [
					{
						"jwtAuth": []
					}
				],
				"parameters": []
			}
		},
...
		"/api/consultations": {
			"get": {
				"operationId": "GetAll",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IPaged_IConsultationResponse_"
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [
					{
						"jwtAuth": []
					}
				],
				"parameters": []
			},
			"post": {
				"operationId": "Create",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IConsultation"
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [
					{
						"jwtAuth": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IConsultationRequest"
							}
						}
					}
				}
			}
		},
		"/api/consultations/{id}": {
			"get": {
				"operationId": "GetOne",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IConsultationDTO"
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [
					{
						"jwtAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "Update",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "boolean"
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [
					{
						"jwtAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IConsultationRequest"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "Delete",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "number",
									"format": "double"
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [
					{
						"jwtAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/consultations/all": {
			"get": {
				"operationId": "All",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/IConsultationDTO"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [
					{
						"jwtAuth": []
					}
				],
				"parameters": []
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}